% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_auto.R
\name{bin.auto}
\alias{bin.auto}
\title{Bin numeric variables using \code{bin.rpart} and \code{bin.custom} automatically}
\usage{
bin.auto(data, x_num = NULL, y)
}
\arguments{
\item{data}{The data frame used for binning}

\item{x_num}{The list of numeric variables. If it is NULL, then all variables except y 
are used.}

\item{y}{The binary outputs}
}
\value{
a list containing cut points and bins
}
\description{
\code{bin.rpart} is tried first. If no bin is found, then \code{bin.custom} is 
used to bin the variable using the cut pionts of 20%, 40%, 60%, and 80% quantiles.
}
\details{
\code{bin.auto} automatically bin a list of numeric variables, and return the 
cut points and bins
}
\examples{
data <- rpart::stagec
x_num <- c('age', 'pgtime')
bin.auto(data, x_num = x_num, y = 'pgstat')
}
