% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_custom.R
\name{bin.custom}
\alias{bin.custom}
\title{Generate bins based on given cut points}
\usage{
bin.custom(x, cut.p, names = NULL, name.style = c("report",
  "scientific"))
}
\arguments{
\item{x}{The numerical data needs to be cut}

\item{cut.p}{The cut poins used for the cutting}

\item{names}{The names for the corresponding bins The poit size for the dots}

\item{name.style}{Either "report" or "scientific" style. The "report" style is 
easier easier to understand for the non-technical audience. While the 
"scientific" style is more succinct.}
}
\value{
Bins the original values fall within
}
\description{
bin.custom() cuts numerical data into different bins based on the given cut 
points. It also gives you the flexibility of naming the bins.
}
\examples{
bin.custom(1:100, cut.p = c(20, 50, 80))
}
