% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_rpart.R
\name{bin.rpart}
\alias{bin.rpart}
\title{Bin numerical values based on recursive partitioning (rpart)}
\usage{
bin.rpart(formula, data, rcontrol = rpart.control(), n.group = NULL, ...)
}
\arguments{
\item{formula}{The formula for rpart}

\item{data}{The data frame used for binning}

\item{n.group}{Number of binned groups}

\item{min.bucket}{The minimum proportion of population in the buckets
(a value between 0 and 1)}
}
\value{
Shows a ggplot with the regression coefficients and the binned groups
}
\description{
\code{bin.rpart} relies on \code{\link[rpart]{rpart}}
function to split numerical values into different nodes. According to the 
tree-structure splits generated form \code{rpart}, \code{bin.rpart} further 
divides the numerical values into different bins. 
The usage of \code{bin.rpart} is similar to \code{rpart}, except that the 
argument of \emph{control} in \code{rpart} is named as \emph{rcontrol} in 
\code{bin.rpart}
}
\examples{
data <- rpart::stagec
bin.knn(pgstat ~ age, data = data, n.group = 4, min.bucket = .1)
# can be combine with the manipulate::manipulate function to change the 
# binning interactively
library(manipulate)
manipulate(bin.knn(pgstat ~ age, data = data, n.group, min.bucket),
  n.group = slider(1, 10, step = 1, initial = 5, label = 'Number of groups'),
  min.bucket = slider(0.01, 1, step = 0.01, initial = 0.05,
  label = 'Minimum Population Size (\%)'))
}

