% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\docType{data}
\name{colorBlindBlack8}
\alias{colorBlindBlack8}
\title{Colorblind friendly palette (black)}
\format{
An object of class \code{character} of length 8.
}
\usage{
colorBlindBlack8
}
\description{
Wong, Bang. ???Color Blindness???. Nature Methods 8, no. 6 (June 2011): 441. https://doi.org/10.1038/nmeth.1618.
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species))+
 geom_point()+
 publish(border = TRUE)+
 scale_color_manual(values = colorBlindBlack8)
}
\keyword{datasets}
