% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorschemes.R
\docType{data}
\name{colorBlindBlack8}
\alias{colorBlindBlack8}
\alias{colorBlindGrey8}
\title{Colorblind friendly palette (grey)}
\format{
An object of class \code{character} of length 8.
}
\usage{
colorBlindGrey8
}
\description{
Colorblind friendly palette (grey)
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species))+
 geom_point()+
 publish(border = TRUE)+
 scale_color_manual(values = colorBlindGrey8)
}
\keyword{datasets}
