% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_col.R
\name{display.col}
\alias{display.col}
\title{Display the color options for ggplot}
\usage{
display.col()
}
\value{
A figure with colors and the corresponding names
}
\description{
The code is based on http://sape.inf.usi.ch/quick-reference/ggplot2/colour, 
with slight pattern change.
}
\examples{
display.col()
}

