% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_pch.R
\name{display.pch}
\alias{display.pch}
\title{Display some frequently-used point shapes for ggplot}
\usage{
display.pch()
}
\value{
A figure with point shapes and corresponding index
}
\description{
The code is based on http://sape.inf.usi.ch/quick-reference/ggplot2/shape, 
with slight pattern change.
}
\examples{
display.pch()
}

