% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.5PL.R
\name{fit.5PL}
\alias{fit.5PL}
\title{Fit a 5 parameter logistical model to dose-response data.}
\usage{
fit.5PL(standards, samples, Rmin = NA, Rmax = NA, n = NA, summary = TRUE)
}
\arguments{
\item{standards}{data.frame; With columns "dose" and "response".}

\item{samples}{data.frame; With column "response".}

\item{Rmin}{numeric; Optional, define the minimum.}

\item{Rmax}{numeric; Optional, define the maximum}

\item{summary}{boolean; Print fit summary, default = TRUE.}
}
\value{
Returns a data.frame with the estimated dose values corresponding to
        the samples response values.
}
\description{
Fit a 5 parameter logistical model to dose-response data.
}
\examples{

df <- dplyr::rename(standard,
                    response = absorbance,
                    dose = concentration)
df2 <- dplyr::rename(samples,
                     response = absorbance.av)
fit.5PL(df, df2, Amin = 0, summary = T)

}
