% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggcorr}
\alias{ggcorr}
\title{Generate the ggplot for a correlation matrix}
\usage{
ggcorr(cor.mat, lower = FALSE, psize = NULL, high = "red", low = "blue",
  digits = 2, var.position = c("axis", "diagonal"), var.angle = 30,
  add.legend = TRUE)
}
\arguments{
\item{cor.mat}{A correlation matrix, usually generated with cor() function}

\item{lower}{Whether to plot the lower triangle only}

\item{psize}{The poit size for the dots}

\item{high}{The color represents the high (positive) correlation}

\item{low}{The color represents the low (negative) correlation}

\item{digits}{The number of digits shown for the correlation values}

\item{var.position}{The position to put the variable names. 
If the value is "axis", the variable names will be at the x and y axis. 
If the value is "diagonal", the variable names will be put along the diagonal
line.}

\item{var.angle}{The angle of the text for the variable names on the X axis or
The diagonal line.}

\item{add.legend}{Whether to add the legend for color (TRUE/FALSE)}
}
\value{
A ggplot of the correlation matrix
}
\description{
Generate the ggplot for a correlation matrix
}
\examples{
xx <- matrix(runif(100), 10)
colnames(xx) <- paste0('Variable ', 1:ncol(xx))
cor.mat <- cor(xx)

ggcorr(cor.mat)  # default plot
ggcorr(cor.mat, var.position = 'diagonal', high = 'blue', low = 'green',
  add.legend = FALSE)
ggcorr(cor.mat, lower = TRUE, var.position = 'diagonal')

}

