% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdistr.R
\name{ggdistr}
\alias{ggdistr}
\title{Plot the distribution (geom_histogram) of each column in the given data frame. 
If the data is categorical data (factor or character), the statistics is set 
as "count". Otherwise, it is set as "bin".}
\usage{
ggdistr(data, n_sample = NULL, ncol = 3, sample_seed = 123, ...)
}
\arguments{
\item{data}{The data frame used for plot}

\item{n_sample}{Number of samples used for the plot. If it is NULL, then all 
records are used}

\item{ncol}{Number of columns for the plot layout}

\item{sample_seed}{The seed used for sampling}
}
\value{
ggplot of the histogram/bar for all variables
}
\description{
\code{ggdistr} plot the data distribution (histogram/bar for numeric/categorical variables)
}
\examples{
data <- rpart::stagec
ggdistr(data)
}

