% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstat.R
\name{ggstat}
\alias{ggstat}
\title{Generate a ggplot for the statistics in different levels (groups)}
\usage{
ggstat(data, var = "Variable.IV", x = "Group", y = "Rate.1",
  y.label = "Perc.1", y.label.col = "red", y.title = NULL,
  bar.col = "cornflowerblue", width = "Rate.group",
  width.label = "Perc.group", width.label.col = "black", ncol = NULL,
  theme = c("classic", "ws"), background = "white")
}
\arguments{
\item{data}{The data frame used for the ggplot. It is usually the output from 
\code{level.stat}. It can be other data frame, if the structure 
of the data is appropriate.}

\item{var}{The column represents the variables, which is used as the head in
each plot panel.}

\item{x}{The variable used as x axis}

\item{y}{The variable used as y axis}

\item{y.label}{The text used to label the y value}

\item{y.label.col}{The color of the \emph{y.label} text}

\item{y.title}{The title of y axis for the whole plot}

\item{bar.col}{The color of the bar}

\item{width}{The column used to represents the width of the bar, or a contant
numeric value.}

\item{width.label}{The text used to label the Width of the bar}

\item{width.label.col}{The color of the \emph{width.label} text}

\item{ncol}{Number of column for the plot}

\item{theme}{The theme of the plot. It can be 'classic' or 'ws' (Wall Street)}

\item{background}{The background color}
}
\value{
a data frame inclduing frequency, rates, Weight of Evidence (WOE), 
   and Information Value (IV) for each level. The return data is already
   ranked by the IV values from high to low.
}
\description{
\code{level.stat} provides the useful statistics for different levels (groups) 
in the given independent variables.
}
\examples{
data <- rpart::stagec
st <- level.stat(data, x = c('eet', 'ploidy'), y = 'pgstat')
ggstat(st, theme = 'ws')
}

