% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importChromatograms.R
\name{importChromatograms}
\alias{importChromatograms}
\title{importChromatograms}
\usage{
importChromatograms(rawfile)
}
\arguments{
\item{rawfile}{file path; File path of an Orbitrap raw file.}
}
\value{
Returns a tibble containing the chromatography information.
}
\description{
importChromatograms
}
\examples{

rawfile <- "C:/myfile.raw"
lc <- importChromatograms(rawfile)

}
