% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_stat.R
\name{level.stat}
\alias{level.stat}
\title{Statistics for binary outputs in different levels and variables}
\usage{
level.stat(data, x = NULL, y, flag.0 = 0, flag.1 = 1)
}
\arguments{
\item{data}{The data frame used for the statistics}

\item{x}{The independent variable. If it is NULL, then all variables except y 
are used.}

\item{y}{The binary outputs}

\item{flag.0}{The value representing 0 in y}

\item{flag.1}{The value representing 1 in y}
}
\value{
a data frame inclduing frequency, rates, Weight of Evidence (WOE), 
   and Information Value (IV) for each level. The return data is already
   ranked by the IV values from high to low.
}
\description{
\code{level.stat} provides the useful statistics for different levels (groups) 
in the given independent variables.
}
\examples{
data <- rpart::stagec
level.stat(data, x = c('eet', 'ploidy'), y = 'pgstat')
}

