% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_auc.R
\name{perf.auc}
\alias{perf.auc}
\title{Check GLM or survival model performance based on AUC}
\usage{
perf.auc(model, train, test)
}
\arguments{
\item{model}{The GLM or survival model (coxph) object.}

\item{train}{The training dataset}

\item{test}{The test dataset}
}
\value{
For a GLM model, it returns the ROC curve and the AUC values for
   the training and test datasets respectively. For a survival model, it 
   returns the time-dependent AUC at each time step and the integrated 
   AUC value (iAUC)
}
\description{
\code{perf.auc} checks model performance for a given model object, 
using the training and test dataset.
}
\examples{
data <- rpart::stagec
data <- data[sample(nrow(data), 10000, replace = TRUE), ]
data <- na.omit(data)

ind.train <- caret::createDataPartition(data$pgstat, p = .7, list = FALSE)
dt.train <- data[ind.train, ]
dt.test <- data[-ind.train, ]
mod <- glm(pgstat ~ ., dt.train, family=binomial(link='logit'))
perf.auc(mod, dt.train, dt.test)

}
