% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_decile.R
\name{perf.decile}
\alias{perf.decile}
\title{Check model performance based on predicated and actual rates}
\usage{
perf.decile(actual, pred, plot = TRUE, add.legend = TRUE)
}
\arguments{
\item{actual}{A vector containing the actual status for each record}

\item{pred}{The predicted probability for each record}

\item{plot}{Whether to show the ggplot figure}

\item{add.legent}{Whether to add a legend for the decile color}
}
\value{
The predicted and actual rates in each decile and a ggplot
}
\description{
\code{perf.decile} takes the actual status (actual), and the predicted 
probability (pred) as inputs, divided the data into 10 decile dependent on 
the ranking of the predicted values, and calculate the average predicted and
actual rates in each decile.
}
\examples{
data <- rpart::stagec
data <- data[sample(nrow(data), 10000, replace = TRUE), ]
data <- na.omit(data)

ind.train <- caret::createDataPartition(data$pgstat, p = .7, list = FALSE)
dt.train <- data[ind.train, ]
dt.test <- data[-ind.train, ]
mod <- glm(pgstat ~ ., dt.train, family=binomial(link='logit'))

pred.test <- predict(mod, newdata = dt.test, type = 'response')
perf.decile(actual = dt.test$pgstat, pred = pred.test)
}
