% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatograms.R
\name{plotChromatograms}
\alias{plotChromatograms}
\title{plotChromatograms}
\usage{
plotChromatograms(lc)
}
\arguments{
\item{lc}{tibble; Output from importChromatograms.}
}
\value{
Returns a ggplot object.
}
\description{
plotChromatograms
}
\examples{

p1 <- plotChromatograms(lc)
p1
ggsave("chromatograms.png", p1)

}
