% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publication ready theme}
\usage{
publish(
  base_size = 12,
  base_family = "sans",
  border = TRUE,
  margin = TRUE,
  legend = c("right", "top", "bottom", "left", "none"),
  x.text.angle = 0,
  flip = FALSE,
  major_grid = TRUE,
  minor_grid = FALSE,
  facet_fill = "grey75",
  facet_color = NA,
  facet_text_color = "white",
  base_line_size = 0.5,
  base_rect_size = 1,
  aspect.ratio = 1
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{border}{logical value. Default is TRUE. If TRUE, add panel border.}

\item{margin}{logical value. Default is TRUE. If FALSE, reduce plot margin.}

\item{legend}{character specifying legend position. Allowed values are one of
c("top", "bottom", "left", "right", "none"). Default is "top" side position.
to remove the legend use legend = "none". Legend position can be also
specified using a numeric vector c(x, y).  In this case it is possible to
position the legend inside the plotting area. x and y are the coordinates of
the legend box. Their values should be between 0 and 1. c(0,0) corresponds
to the "bottom left" and c(1,1) corresponds to the "top right" position. For
instance use legend = c(0.8, 0.2).}

\item{x.text.angle}{Rotation angle of x axis tick labels. Default value is 0.
Use 90 for vertical text.}

\item{flip}{logical. If TRUE, grid lines are added to y axis instead of x
axis.}

\item{major_grid}{logical. If TRUE, major grid lines are added.}

\item{minor_grid}{logical. If TRUE, minor grid lines are added.}

\item{facet_fill}{fill color for facet strips. Defaults to "grey75".}

\item{facet_color}{border color for facet strips. Defaults to "NA".}

\item{facet_text_color}{text color for facet strips. Defaults to "white".}

\item{base_line_size}{numerical. Specifies the default element_line() size
for plots. Defaults to 0.5.}

\item{base_rect_size}{numerical. Specifies the default line size for
element_rect() in plots. Defaults to 1.}

\item{aspect.ratio}{numerical. Specifies the aspect ratio of the plot.}
}
\description{
Create a publication ready theme based on the ggpubr() package.
The publish() function is a customized version of theme_bw() adjusted to my
personal preference.
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(aes(color = gear))

# Default plot
p

# Use publish()
p + publish()

}
