% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_woe.R
\name{replace.woe}
\alias{replace.woe}
\title{replace original levels with Weight of Evidence (WOE)}
\usage{
replace.woe(data, stat, replace = FALSE)
}
\arguments{
\item{data}{The data frame needs to be replaced}

\item{stat}{The statistics generated from \code{level.stat}}

\item{replace}{Whether to replace (replace = TRUE) the original values or 
add a new column to the end of the data (replace = FALSE).}
}
\value{
a data frame with WOE values.
}
\description{
\code{replace.woe} replace the original levels with the calculated WOE from 
the \code{level.stat} function. It will replace all variables that can be 
found from \emph{stat}
}
\examples{
data <- rpart::stagec
st <- level.stat(data, x = c('eet', 'ploidy'), y = 'pgstat')
replace.woe(data, stat = st)
}

