% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_df.R
\name{summary_df}
\alias{summary_df}
\title{Get the summary of data, which includes Column_Name, Data_Type, 
N_Unique (Number of unique values), Values (range for non-factor and non-character,
and categories for factor and character), and NA_Ratio. The result is returned 
as a data frame.}
\usage{
summary_df(data)
}
\arguments{
\item{data}{The data frame to be summarized}
}
\description{
\code{summary_df} provides the useful statistics for all columns
}
\examples{
data <- rpart::stagec
summary_df(data)
}
