% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{a classic simple theme with x and y axis}
\usage{
theme_simple(...)
}
\value{
A ggplot with simple theme
}
\description{
a classic simple theme with x and y axis
}
\examples{
ggplot(mtcars, aes(x = mpg, y = disp, color = as.factor(cyl))) + 
  geom_point() + 
  theme_simple()
}
