% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ws.R
\name{theme_ws}
\alias{theme_ws}
\title{A theme like Wall Street Journal figures}
\usage{
theme_ws(background = "ivory", ...)
}
\value{
A ggplot with Wall-Street-Journal theme
}
\description{
A theme for ggplot, which looks like the Wall Street Journal figures. 
It is similar to the \emph{theme_wsj} in the \emph{ggthemes} package, with the
font simplified.
}
\examples{
ggplot(mtcars, aes(x = mpg, y = disp, color = as.factor(cyl))) + 
  geom_point() + 
  theme_ws()
}
