% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotToGenename.R
\name{UniprotToGenename}
\alias{UniprotToGenename}
\title{UniprotToGenename}
\usage{
UniprotToGenename(ids)
}
\arguments{
\item{ids}{A character vector of Uniprot IDs.}
}
\description{
This is a simple function which uses the Uniprot REST API to
request the gene names from a list of uniprot IDs.
}
\examples{
contaminants <- c("P19013", "Q7RTT2", "P15636", "P09870", "Q9R4J5")
UniprotToGenename(ids = contaminants)

}
