% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_rpart.R
\name{bin.rpart}
\alias{bin.rpart}
\title{Bin numerical values based on recursive partitioning (rpart)}
\usage{
bin.rpart(formula, data, rcontrol = rpart.control(), n.group = NULL,
  ...)
}
\arguments{
\item{formula}{The formula for rpart}

\item{data}{The data frame used for binning}

\item{rcontrol}{The arguments passed into \code{\link[rpart]{rpart.control}}}

\item{n.group}{Number of acceptable binning groups. It can be NULL,
a single number (e.g., 5), or a vector (e.g., 3:7). If the value is NULL, it 
returns the output with the default \emph{rpart.control}
settings. If the n.group is a numeric value, it will change the \emph{cp} 
value within \emph{rpart.control} automatically, 
until it gets the desirable number of groups}

\item{...}{All other arguments that can be passed to \code{rpart}}
}
\value{
The cut points (\emph{cut.points}) and \emph{bins}.
}
\description{
\code{bin.rpart} relies on \code{\link[rpart]{rpart}}
function to split numerical values into different nodes. According to the 
tree-structure splits generated form \code{rpart}, \code{bin.rpart} further 
divides the numerical values into different bins, and record the cut points. 
The usage of \code{bin.rpart} is similar to \code{rpart}, except that the 
argument of \emph{control} in \code{rpart} is named as \emph{rcontrol} in 
\code{bin.rpart}
}
\examples{
data <- rpart::stagec
bin.rpart(pgstat ~ age, data = data)
}
